Basic Layout before implementing router
