--- exemples/TabSolutionsClient/src/app/add-new/add-new.component.html
@@ -1,3 +1,0 @@
-<p>
-  add-new works!
-</p>
