+++ exemples/TabSolutionsClient/src/app/app.module.ts
@@ -1,16 +1,15 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
+import { Routes, RouterModule } from '@angular/router';
 
 import { AppComponent } from './app.component';
 import { MainScreenComponent } from './main-screen/main-screen.component';
-import { AddNewComponent } from './add-new/add-new.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    MainScreenComponent,
-    AddNewComponent
+    MainScreenComponent
   ],
   imports: [
     BrowserModule,
