+++ exemples/TabSolutionsClient/src/index.html
@@ -8,8 +8,7 @@
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
 <body>
-  <app-root></app-root>
-  <app-main-screen></app-main-screen>
-  <div ng-include="'main-screen-component.html'"></div>
+  <app-root>
+  </app-root>
 </body>
 </html>
